move tests to header
