+++ static/app/views/issueDetails/streamline/eventDetails.spec.tsx
@@ -1,5 +1,4 @@
 import {EventFixture} from 'sentry-fixture/event';
-import {EventsStatsFixture} from 'sentry-fixture/events';
 import {GroupFixture} from 'sentry-fixture/group';
 import {LocationFixture} from 'sentry-fixture/locationFixture';
 import {OrganizationFixture} from 'sentry-fixture/organization';
@@ -7,7 +6,7 @@ import {ProjectFixture} from 'sentry-fixture/project';
 import {RouterFixture} from 'sentry-fixture/routerFixture';
 import {TagsFixture} from 'sentry-fixture/tags';
 
-import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
+import {render, screen} from 'sentry-test/reactTestingLibrary';
 
 import PageFiltersStore from 'sentry/stores/pageFiltersStore';
 import ProjectsStore from 'sentry/stores/projectsStore';
@@ -15,7 +14,6 @@ import {EventDetails} from 'sentry/views/issueDetails/streamline/eventDetails';
 import {MOCK_EVENTS_TABLE_DATA} from 'sentry/views/performance/transactionSummary/transactionEvents/testUtils';
 
 jest.mock('sentry/views/issueDetails/groupEventDetails/groupEventDetailsContent');
-jest.mock('sentry/views/issueDetails/streamline/issueContent');
 jest.mock('screenfull', () => ({
   enabled: true,
   isFullscreen: false,
@@ -23,11 +21,6 @@ jest.mock('screenfull', () => ({
   exit: jest.fn(),
   on: jest.fn(),
   off: jest.fn(),
-}));
-
-const mockUseNavigate = jest.fn();
-jest.mock('sentry/utils/useNavigate', () => ({
-  useNavigate: () => mockUseNavigate,
 }));
 
 describe('EventDetails', function () {
@@ -38,8 +31,6 @@ describe('EventDetails', function () {
   const defaultProps = {project, group, event};
 
   let mockActionableItems: jest.Mock;
-  let mockTags: jest.Mock;
-  let mockStats: jest.Mock;
   let mockList: jest.Mock;
   let mockListMeta: jest.Mock;
 
@@ -60,18 +51,11 @@ describe('EventDetails', function () {
       body: {errors: []},
       method: 'GET',
     });
-    mockTags = MockApiClient.addMockResponse({
+    MockApiClient.addMockResponse({
       url: `/organizations/${organization.slug}/issues/${group.id}/tags/`,
       body: TagsFixture(),
       method: 'GET',
     });
-
-    mockStats = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events-stats/`,
-      body: {'count()': EventsStatsFixture(), 'count_unique(user)': EventsStatsFixture()},
-      method: 'GET',
-    });
-
     mockList = MockApiClient.addMockResponse({
       url: '/organizations/org-slug/events/',
       headers: {
@@ -111,14 +95,6 @@ describe('EventDetails', function () {
     render(<EventDetails {...defaultProps} />, {organization});
     await screen.findByText(event.id);
 
-    // Filtering
-    expect(mockTags).toHaveBeenCalled();
-    expect(screen.getByTestId('page-filter-environment-selector')).toBeInTheDocument();
-    expect(screen.getByLabelText('Search events')).toBeInTheDocument();
-    expect(screen.getByTestId('page-filter-timerange-selector')).toBeInTheDocument();
-    // Graph
-    expect(mockStats).toHaveBeenCalled();
-    expect(screen.getByRole('figure')).toBeInTheDocument();
     // Navigation
     expect(screen.getByRole('tab', {name: 'Recommended Event'})).toBeInTheDocument();
     expect(screen.getByRole('tab', {name: 'First Event'})).toBeInTheDocument();
@@ -145,53 +121,5 @@ describe('EventDetails', function () {
 
     expect(mockList).toHaveBeenCalled();
     expect(mockListMeta).toHaveBeenCalled();
-  });
-
-  it('displays error messages from bad queries', async function () {
-    const errorMessage = 'wrong, try again';
-    mockStats = MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/events-stats/`,
-      body: {detail: errorMessage},
-      method: 'GET',
-      statusCode: 400,
-    });
-
-    render(<EventDetails {...defaultProps} />, {organization});
-    await screen.findByText(event.id);
-
-    expect(mockStats).toHaveBeenCalled();
-    expect(screen.getByText(errorMessage)).toBeInTheDocument();
-    // Omit the graph
-    expect(screen.queryByRole('figure')).not.toBeInTheDocument();
-  });
-
-  it('updates the query params with search tokens', async function () {
-    const [tagKey, tagValue] = ['user.email', 'leander.rodrigues@sentry.io'];
-    const locationQuery = {
-      query: {
-        query: `${tagKey}:${tagValue}`,
-      },
-    };
-    MockApiClient.addMockResponse({
-      url: `/organizations/${organization.slug}/tags/${tagKey}/values/`,
-      body: [
-        {
-          key: tagKey,
-          name: tagValue,
-          value: tagValue,
-        },
-      ],
-      method: 'GET',
-    });
-
-    render(<EventDetails {...defaultProps} />, {organization});
-    await screen.findByText(event.id);
-
-    const search = screen.getAllByRole('combobox', {name: 'Add a search term'})[0];
-    await userEvent.type(search, `${tagKey}:`);
-    await userEvent.keyboard(`${tagValue}{enter}{enter}`);
-    expect(mockUseNavigate).toHaveBeenCalledWith(expect.objectContaining(locationQuery), {
-      replace: true,
-    });
   });
 });
